Refs: https://github.com/MARKETProtocol/dApp/pull/86#issuecomment-380041739
